client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "httpx>=0.28.1",
    "pillow>=12.1.1",
    "pymongo>=4.16.0",
    "pysocks>=1.7.1",
    "python-dotenv>=1.2.2",
    "reportlab>=4.4.10",
    "socks>=0",
]
